Naomi: allow mapping P2 buttons onto P1. Fix coin chute issues.

+++ core/hw/maple/maple_cfg.cpp
@@ -23,7 +23,7 @@ Plugins:
 void UpdateInputState(u32 port);
 void UpdateVibration(u32 port, u32 value, u32 max_duration);
 
-extern u16 kcode[4];
+extern u32 kcode[4];
 extern u32 vks[4];
 extern s8 joyx[4],joyy[4];
 extern u8 rt[4],lt[4];
